+++ lib/integrations/api_exceptions.rb
